add action to add words to forq
