+++ src/shared/jpdb/add-vocabulary.ts
@@ -1,5 +1,6 @@
 import { JPDBAddVocabularyRequest, JPDBRequestOptions } from './api.types';
 import { request } from './request';
+import { requestByUrl } from './request-by-url';
 import { JPDBSpecialDeckNames } from './types';
 
 export const addVocabulary = async (
@@ -12,6 +13,19 @@ export const addVocabulary = async (
     id,
     vocabulary: [[vid, sid]],
   };
+
+  if (id === 'forq') {
+    return await requestByUrl(
+      undefined,
+      'prioritize',
+      {
+        v: vid,
+        s: sid,
+        origin: '/',
+      },
+      options,
+    );
+  }
 
   await request('deck/add-vocabulary', payload, options);
 };
