+++ src/shared/jpdb/api.types.ts
@@ -74,4 +74,5 @@ export type JPDBEndpoints = {
   'list-user-decks': [JPDBListUserDecksRequest, JPDBListUserDecksResult];
   'deck/add-vocabulary': [JPDBAddVocabularyRequest, void];
   'deck/remove-vocabulary': [JPDBRemoveVocabularyRequest, void];
+  prioritize: [{ v: number; s: number; origin: string }, void];
 };
